Do not rely on VIRTUAL_ENV in elpa-get
In tox, $VIRTUAL_ENV is not set.  So, running tox in virtualenv
causes trouble because elpa-get tries to find emacs executable in
the outer virtualenv instead of the one crated by tox.  I propose
to not rely on the environment variable $VIRTUAL_ENV even though
it is not robust to moving virtualenv directory around.  But as
other scripts in virtualenv have hard-coded shebang, there is no
need to make only elpa-get relocation-friendly.